requirements.txt
Flask
Flask-Cors
python-chess
gunicorn